An alias for opening an unsafe Chrome instance
This instance of Chrome will not enforce the Same Origin Policy which
will make debugging without CORS easier.
To ensure it is opened as a separate instance, you must give it a new
user-data directory.
This also means its possible to give this instance its own settings. I
have chosen to give it a bright red skin so that it is very obvious that
browser is not safe!