Refactor bashrc settings, git prompt, etc.
I'm extracting the git_status_prompt function into it's own file so that I could
use it with other shells, such as fish or zsh.
Other shells shouldn't have any problem running this as a script because it has
the bash shebang and should be interpreted by bash regardless of the calling
shell.