Install FISH
This is a bit of an experiment, fish has some nice advantages but requires its
own ecosystem of tools that is different that what I'm used to for dev work.

The previous refactor allows my prompt to look similar in fish and bash. This is
nice because its likely that I'll need to switch back to bash when something
doesn't work properly in fish.